+++ src/dotools_py/settings/__init__.py
@@ -1,1 +1,2 @@
-from dotools_py.settings._configure import interactive_session, session_settings, matplotlib_backend
+from dotools_py.settings._configure import (interactive_session, session_settings, matplotlib_backend,
+                                            set_kernel_logger, toogle_kernel_logger)